[004_wrong_file_load], issue #4, Stop wrong files loading
In order to prevent the application from loading any file with a wrong extension (different from either '.inp' or '.dat/.plt'), a return statement has been added to the 'load_inp_file' and 'load_output_files' methods of the 'TuPostProcessingGui' class. In this way, when the user closes the pop-up window showing the error message, boths methods exit safely, and no further loading of the selected files is performed.