[030_loading_da_files], issue #30, Updated direct access file names extraction
The 'init_PliReader()' static method of the 'PliReader' class has been updated in the way it searches for the names of the direct access files (i.e. .mic/.mac/.sta) that are specified in the .pli file. Previously, cases where those file names contained a hyphen were not handled properly; now, any letter, digit, underscore or hyphen is admitted and those files are correctly found when reading the input .pli file.