[031_wrong_no_slices_times], issue #31, Corrected slice identification
The 'init_PliReader()' static method of the 'PliReader' class has been updated in the way it extracts the number of slices from the .pli file. This is done by reading the 'ISLICE' value; depending on its value, the number of slices is calculated as 'M3'+1, if 'ISLICE' is 0, or as 'M3', if 'ISLICE' is 1, where the 'M3' value is found in the .pli file as well. The 'ISLICE' is read as a string; however, previously, it was compared with an integer value, thus preventing to correctly identify the number of slices. Now, the 'ISLICE' value is compared with a string, thus solving the identified issue. By properly identifying the number of slices, also the problem on the missing time instants for the radially dependent quantities is solved; in fact, the extraction of these times from the .mac file is directly tied to the number of slices.