[040_no_output], issue #40, Added check on files last modification time
So far, if any problem in the generation of the output plot files happened, no error message was shown to the user if the output folder already had plot files with the same names. This has been addressed by checking the last modification time of the output files: if any change in the last modification time occurs, it means the files have been generated, otherwise an exception is raised and a pop-up message is shown.