[038_limit_curves_selection], issue #38, Limited selection in listbox
The listbox widget, where several curves can be selected to be plotted at once, now has a limited number of selectable items. In the 'PlotSettingsListBox', the 'MAX_SELECTABLE_ITEMS' attribute drives how many curves can be plotted. If more items than the admitted one are selected, only the number of items indicated by 'MAX_SELECTABLE_ITEMS' are actually shown as selected and can be plotted. Any further attempt to selected another item in the listbox widget does not produce any result. This limit has been introduced to cope with the fact that the plotting executable can produce a limited number of plot files.