[035_output_folder_info], issue #35, Changed logic for assigning output folder
Previously, if not set directly by the menu bar option, the output folder was assigned to the one of the firstly loaded .pli file. If any other file was opened, this directory did not change anymore. In addition, there was no indication for the user about this behaviour. Now, if not directly set, the output folder is always assigned to the directory of the currently loaded .pli or .inp file. In the status bar, the path of the loaded file was displayed; now, also the path of the output folder is shown.