[035_output_folder_info], issue #35, Reset output folder
Resetting the GUI layout using the 'File->New' command now clears any previously set path for the output folder. The assignment of this folder will then follow the currently implemented behaviour: the folder will be set to the path of the directory of the .pli or .inp file that is subsequently loaded, if not set directly.